source/client/requirements.txt
paho-mqtt==1.4.0
PyYAML==5.1
